Added really basic anonymous accounts.
Essentially, I added backend functionality for providing the ability to access the site anonymously. It still generates an account, just one that a user cannot necessarily access again very easy. It generates a random username, emails are now optional for he sake of making this easier on us. In the login component there is a button to access anonymously.